# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API UHS Asbl",
        "description": "API REST complète avec gestion des activites de UHS Asbl",
        "contact": {
            "name": "Support API",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-uhs.uhs-rdc.org",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/transaction-tresoreries.store": {
            "post": {
                "tags": [
                    "TransactionTrésoreries"
                ],
                "summary": "Créer une transaction de trésorerie",
                "operationId": "b944dc16a8d315859c6a23613b536ba5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_date",
                                    "account_id",
                                    "amount",
                                    "transaction_type"
                                ],
                                "properties": {
                                    "motif": {
                                        "type": "string",
                                        "example": "Achat fournitures"
                                    },
                                    "transaction_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-27"
                                    },
                                    "account_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 500
                                    },
                                    "transaction_type": {
                                        "type": "string",
                                        "example": "depot"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction créée avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    },
                    "409": {
                        "description": "Transaction déjà existante"
                    }
                }
            }
        },
        "/api/transaction-tresoreries.update/{id}": {
            "put": {
                "tags": [
                    "TransactionTrésoreries"
                ],
                "summary": "Mettre à jour une transaction de trésorerie",
                "operationId": "9c76982ae0707a6f9f8fc84cec6f0e24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "motif": {
                                        "type": "string",
                                        "example": "Achat fournitures"
                                    },
                                    "transaction_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-27"
                                    },
                                    "account_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 500
                                    },
                                    "transaction_type": {
                                        "type": "string",
                                        "example": "DEPENSE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction mise à jour avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                }
            }
        },
        "/api/transaction-tresoreries.delete/{id}": {
            "delete": {
                "tags": [
                    "TransactionTrésoreries"
                ],
                "summary": "Supprimer une transaction de trésorerie",
                "operationId": "47045cbfa7c03809f5e31d68cf0f4488",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction supprimée avec succès"
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                }
            }
        },
...
        "/api/user.delete/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a User",
                "description": "Delete a user by their ID.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}